api, upload: move upload package code to api package
